requirements.txt
fastapi>=0.95.0,<0.96.0
uvicorn>=0.22.0,<0.23.0
sqlalchemy>=2.0.0,<2.1.0
pydantic>=1.9.0,<2.0.0
python-multipart>=0.0.6,<0.1.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
email-validator>=2.0.0,<2.1.0
pytest>=7.3.1,<7.4.0
